+++ CHANGELOG.md
@@ -16,6 +16,10 @@
   (`$HOME/.singularity/docker-config.json`). The commands `pull`, `push`, `run`,
   `exec`, `shell`, and `instance start` can now also be passed a `--authfile
   <path>` option, to read OCI registry credentials from this custom file.
+- A new `--keep-layers` flag, for the `pull` and `run/shell/exec/instance start`
+  commands, allows individual layers to be preserved when an OCI-SIF image is
+  created from an OCI source. Multi layer OCI-SIF images can be run with
+  SingularityCE 4.1 and later.
 
 ### Bug Fixes